code for a data assignment
To design efficient materials for organic photovoltaics (OPVs) the most important parameter is the power conversion efficiency (PCE)  of a material. Here is a dataset of 280 samples taken from the article Adv. Energy Mater. 2018, 1801032 by Sahu et al. Each of the sample/material has 13 independent features and it belongs to a class depending on its PCE value (low, moderate, high). Construct ML classifiers using logistic regression, support vector machine, and random forest algorithms to classify the samples. Report the accuracies for the training and test sets. Find out the top four important features. The article and the datasets (.xlsx files) are provided here.

In two more files (Train.csv and Test.csv), the values for PCE are given along with all other independent features. Build a regression model using a multiple linear regression algorithm and calculate the RMSE and R2 for the training and test datasets.